+++ src/groovy/net/hedtech/restfulapi/marshallers/xml/DeclarativeBeanMarshaller.groovy
@@ -50,6 +50,7 @@ class DeclarativeBeanMarshaller extends BeanMarshaller {
     def marshalledNullFields = [:]
     //default behavior on whether to marshall null fields.
     def marshallNullFields = true
+    def marshallEmptyCollections = true
 
 
     @Override
@@ -108,6 +109,9 @@ class DeclarativeBeanMarshaller extends BeanMarshaller {
 
         if (ignoreNull) {
             Object val = beanWrapper.getPropertyValue(property.getName())
+            if (!marshallEmptyCollections && val instanceof Collection && val.size() == 0) {
+                return false
+            }
             return val != null
         } else {
             return true
@@ -127,6 +131,9 @@ class DeclarativeBeanMarshaller extends BeanMarshaller {
 
         if (ignoreNull) {
             Object val = field.get(value)
+            if (!marshallEmptyCollections && val instanceof Collection && val.size() == 0) {
+                return false
+            }
             return val != null
         } else {
             return true