+++ src/groovy/net/hedtech/restfulapi/marshallers/xml/DeclarativeDomainClassMarshaller.groovy
@@ -48,6 +48,7 @@ class DeclarativeDomainClassMarshaller extends BasicDomainClassMarshaller {
     def marshalledNullFields = [:]
     //default behavior on whether to marshall null fields.
     def marshallNullFields = true
+    def marshallEmptyCollections = true
 
     def additionalFieldClosures = []
     def additionalFieldsMap = [:]
@@ -169,6 +170,9 @@ class DeclarativeDomainClassMarshaller extends BasicDomainClassMarshaller {
 
         if (ignoreNull) {
             Object val = beanWrapper.getPropertyValue(property.getName())
+            if (!marshallEmptyCollections && val instanceof Collection && val.size() == 0) {
+                return false
+            }
             return val != null
         } else {
             return true